+++ app/helpers/application_helper.rb
@@ -1899,20 +1899,4 @@ module ApplicationHelper
       wiki_pages: auto_complete_wiki_pages_path(project_id: project, q: ''),
     }
   end
-
-  def minimum_time(current_user)
-    return nil unless Setting.allow_logging_time.eql?('1') || !current_user.admin
-
-    if Time.current.monday?
-      if Time.current.hour.to_i >= Setting.allow_logging_time_till.to_time.hour.to_i
-        Date.current.strftime('%Y-%m-%d')
-      else
-        prior_friday(Date.current)
-      end
-    elsif Time.current.hour.to_i >= Setting.allow_logging_time_till.to_time.hour.to_i
-      Date.current.strftime('%Y-%m-%d')
-    else
-      Date.yesterday.strftime('%Y-%m-%d')
-    end
-  end
 end
