+++ plugins/mail_tracker/app/helpers/application_helper_patch.rb
