+++ plugins/mail_tracker/init.rb
@@ -25,4 +25,6 @@ Proc.new do
   ReportsController.send(:include, ReportsControllerPatch)
   WatchersController.send(:include, WatchersControllerPatch)
   ApplicationHelper.send(:include, ApplicationHelperPatch)
+  SettingsHelper.send(:include, SettingsHelperPatch)
+  Setting.send(:include, SettingPatch)
 end.call
